hs: Add a flag to export circuit id per port
Previously we have HiddenServiceExportCircuitID which allows us to
export the circuit id using haproxy. However, such directive is applied
globally in the service. We need a way to export the circuit id only for
some virtual ports.

We can do that by adding an optional flag to the HiddenServicePort
directive to specify that we will export the circuit id for such port.
The flag in HiddenServicePort will override the one in
HiddenServiceExportCircuitID, if it's not none. But it will use the
protocol specified in HiddenServiceExportCircuitID, if it's otherwise.